pyproject.toml
[project]
name = "diarization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "pyannote-audio==3.4.0",
    "huggingface-hub<1.0.0",
    "soundfile",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "torchaudio>=2.6.0",
    "openai-whisper>=20250625",
    "transformers>=4.57.1",
    "accelerate>=1.11.0",
    "bitsandbytes>=0.48.1",
    "compressed-tensors>=0.15.0.1",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
